backend/apps/catalog/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from . import views

router = DefaultRouter(trailing_slash=False)
router.register('categories', views.CategoryViewSet, basename='category')
router.register('items', views.ItemViewSet, basename='item')

urlpatterns = [
    path('/', include(router.urls)),
]
