backend/requirements.txt
Django==4.2.*
djangorestframework==3.15.*
django-cors-headers==4.4.*
psycopg2-binary==2.9.*
celery==5.4.*
redis==5.0.*
anthropic==0.39.*
httpx==0.27.*
python-decouple==3.8
dj-database-url==2.2.*
gunicorn==22.0.*
